Show help and guide documents integrated in the website
Before when one of the docs was opened, all gnucash context website was gone.
There is one visible artifact that may still need some attention. The iframe
used to display the document is fixed size and hence can sometimes show a
second scrollbar. This can't be fixed with html/css alone.

For more details on how this grew, please read the commit history of the beta
branch (r22480-r22507).

git-svn-id: svn+ssh://svn.gnucash.org/repo/htdocs/trunk@22520 57a11ea4-9604-0410-9ed3-97b8803252fd