Update header.phtml
To improve the multilinguality of the website in web-search results, I propose some changes in the meta tags of the <head> of the „header.phtml” file. The idea is to use translated text for the other languages.

Changes would be made for:
a) meta property="og:title"
b) meta property="og:description"
c) meta name="description"

It has something to do with php stuff. As I’m not a programmer, I’m not sure if this solution works, nor if it is legit. But something similar is being used in the „menu.phtml” file.
!!! Please see the diff between my new file and the orig file on master !!!

P.S.
I’ve checked my „experiment” at https://metatags.io/
It seems that html-tags within then strings don’t matter (which I was afraid of).

If my changes don’t work, maybe somebody knows how to handle the issue. But that would probably be something for a feature request ...