Adjust help document loading to accomodate ghelp2help changes in gnucash-docs
The issue to deal with is that the help manual's main page gets renamed from help.html to index.html.
But only for documentation from gnucash 5 and up.
To handle this, delay document selection until after version selection.
Use the version selection to set the proper name for the main page.